Refactor to get rid of #minitest_reporter methods on Queue implementations
