+++ ruby/lib/ci/queue/redis/retry.rb
@@ -11,15 +11,6 @@ module CI
           @build ||= CI::Queue::Redis::BuildRecord.new(self, redis, config)
         end
 
-        def minitest_reporters
-          require 'minitest/reporters/queue_reporter'
-          require 'minitest/reporters/redis_reporter'
-          @minitest_reporters ||= [
-            Minitest::Reporters::QueueReporter.new,
-            Minitest::Reporters::RedisReporter::Worker.new(build: build),
-          ]
-        end
-
         private
 
         attr_reader :redis
